requirements.txt
# Runtime dependencies for Quantum Tutor v6.1
bcrypt>=4.0.0
streamlit>=1.40.0
fastapi
uvicorn
httpx
python-multipart
pandas
sentence-transformers
sympy
qrcode[pil]
requests
PyMuPDF
easyocr
torch
torchvision
numpy
python-dotenv
google-generativeai
# Pinecone connector for production
pinecone-client
# Supabase for persistent learning analytics
supabase
# PWA support handled manually via static/manifest.json
